Remove the `/search` route.
- Make `/` resposible for searching,
- update the tests accordingly.
